refactor: use specific models in routes instead of dbRequests (now deleted)
